Removing the ability of AdaptiveKDTree to have multiple trees per object. Fixing a number of 'narrowing conversion' warnings from newer version of gcc.
Passes all tests, serial and parallel.



git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@5982 6162379e-bd28-0410-9a7d-b7f4dcfcad3c

+++ IODebugTrack.cpp
@@ -104,7 +104,7 @@ IODebugTrack::~IODebugTrack()
 void IODebugTrack::record_io( unsigned long begin, unsigned long count )
 {
   if (enableOutput && count) {
-    DRange ins = { begin, begin+count-1, mpiRank };
+    DRange ins = { begin, begin+count-1, static_cast<long unsigned>(mpiRank) };
     record_io( ins );
   }
 }
